# docs/swagger/gym-platform.yaml
...
openapi: 3.0.3
info:
  title: Gym Management Backend API
  description: |-
    Custom OpenAPI documentation for the Gym Management microservices. The platform
    manages authentication, trainees, trainers, assignments, training sessions, and
    synchronized monthly trainer workload.
  contact:
    name: Gym Management API Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
...
servers:
- url: http://localhost:8080
  description: Gym Management microservice
- url: http://localhost:8080/workload
  description: Trainer Workload microservice
...
paths:
...
  "/api/v1/trainees/update/{username}":
    put:
      tags:
      - Gym Management Trainee Controller
      summary: Update a trainee profile by username
      operationId: updateTraineeProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateTraineeProfileRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserProfileResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainees/update/{username}/trainers":
    put:
      tags:
      - Gym Management Trainee Controller
      summary: Update the list of trainers assigned to a trainee
      operationId: updateTraineeTrainers
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateTraineeTrainersRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/UserSummaryResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/auth/change-password":
    put:
      tags:
      - Gym Management Authentication Controller
      summary: Change the password of the current user
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ChangePasswordRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
...
  "/api/v1/trainees/register":
    post:
      tags:
      - Gym Management Trainee Controller
      summary: Register a new trainee
      operationId: registerTrainee
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TraineeRegistrationRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AuthResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/auth/sign-out":
    post:
      tags:
      - Gym Management Authentication Controller
      summary: Sign out the current user and invalidate the JWT token
      operationId: signOut
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
  "/api/v1/auth/register":
    post:
      tags:
      - Gym Management Authentication Controller
      summary: Register a new user and send an email confirmation code
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RegisterRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AuthResponseDTO"
  "/api/v1/auth/login":
    post:
      tags:
      - Gym Management Authentication Controller
      summary: Authenticate a user and return a JWT token
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AuthResponseDTO"
  "/api/v1/auth/confirmation":
    post:
      tags:
      - Gym Management Authentication Controller
      summary: Confirm email with a code and return a JWT token
      operationId: confirm
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ConfirmRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AuthResponseDTO"
  "/api/v1/auth/confirmation/resend":
    post:
      tags:
      - Gym Management Authentication Controller
      summary: Resend an email confirmation code
      operationId: resendConfirmationCode
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ResendConfirmationRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AuthResponseDTO"
  "/api/v1/admin/register":
    post:
      tags:
      - Gym Management Administration Controller
      summary: Register a new admin
      operationId: registerAdmin
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AdminRegistrationRequestDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
...
  "/api/v1/trainees/{username}/deactivate":
    patch:
      tags:
      - Gym Management Trainee Controller
      summary: Deactivate a trainee profile by username
      operationId: deactivateTraineeProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainees/{username}/activate":
    patch:
      tags:
      - Gym Management Trainee Controller
      summary: Activate a trainee profile by username
      operationId: activateTraineeProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
...
  "/api/v1/trainees/{username}/not-assigned-trainers":
    get:
      tags:
      - Gym Management Trainee Controller
      summary: Get a list of active trainers not assigned to the trainee
      operationId: getNotAssignedActiveTrainers
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/UserSummaryResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/trainees/profile/{username}":
    get:
      tags:
      - Gym Management Trainee Controller
      summary: Get a trainee profile by username
      operationId: getTraineeProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserProfileResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/auth/me":
    get:
      tags:
      - Gym Management Authentication Controller
      summary: Get information about the current authenticated user
      operationId: currentSession
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AuthResponseDTO"
      security:
      - Bearer Authentication: []
  "/api/v1/admin/dashboard":
    get:
      tags:
      - Gym Management Administration Controller
      summary: Get admin dashboard data
      operationId: dashboard
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AdminDashboardResponseDTO"
      security:
      - Bearer Authentication: []
...
  "/api/v1/trainees/delete/{username}":
    delete:
      tags:
      - Gym Management Trainee Controller
      summary: Delete a trainee profile by username
      operationId: deleteTraineeProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MessageResponseDTO"
      security:
      - Bearer Authentication: []
...
components:
  schemas:
    UserProfileResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
        specialization:
          type: string
        profileStatus:
          type: string
          enum:
          - Active
          - Inactive
          - Suspended
          - Pending
          - Deleted
        trainers:
          type: array
          items:
            "$ref": "#/components/schemas/UserSummaryResponseDTO"
        trainees:
          type: array
          items:
            "$ref": "#/components/schemas/UserSummaryResponseDTO"
    UserSummaryResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        specialization:
          type: string
        active:
          type: boolean
        profileStatus:
          type: string
          enum:
          - Active
          - Inactive
          - Suspended
          - Pending
          - Deleted
        role:
          type: string
          enum:
          - Admin
          - Trainee
          - Trainer
...
    UpdateTraineeProfileRequestDTO:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
          minLength: 1
        isActive:
          type: boolean
      required:
      - address
      - dateOfBirth
      - firstName
      - isActive
      - lastName
    UpdateTraineeTrainersRequestDTO:
      type: object
      properties:
        trainers:
          type: array
          items:
            type: string
            minLength: 1
          minItems: 1
      required:
      - trainers
    ChangePasswordRequestDTO:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 1
        oldPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          maxLength: 2147483647
          minLength: 8
      required:
      - email
      - newPassword
      - oldPassword
      - username
    MessageResponseDTO:
      type: object
      properties:
        message:
          type: string
...
    AuthResponseDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        message:
          type: string
        token:
          type: string
        tokenType:
          type: string
        role:
          type: string
    TraineeRegistrationRequestDTO:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
      required:
      - firstName
      - lastName
    RegisterRequestDTO:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        passConfirm:
          type: string
      required:
      - email
      - password
    LoginRequestDTO:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
        password:
          type: string
          minLength: 1
      required:
      - password
    ConfirmRequestDto:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        confirmation:
          type: string
          minLength: 1
      required:
      - confirmation
      - email
    ResendConfirmationRequestDTO:
      type: object
      properties:
        email:
          type: string
          minLength: 1
      required:
      - email
    AdminRegistrationRequestDTO:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 1
        password:
          type: string
      required:
      - firstName
      - lastName
      - username
...
    AdminDashboardResponseDTO:
      type: object
      properties:
        totalUsers:
          type: integer
          format: int64
        activeUsers:
          type: integer
          format: int64
        inactiveUsers:
          type: integer
          format: int64
        admins:
          type: integer
          format: int64
        trainees:
          type: integer
          format: int64
        trainers:
          type: integer
          format: int64
        trainings:
          type: integer
          format: int64
        users:
          type: array
          items:
            "$ref": "#/components/schemas/UserSummaryResponseDTO"
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Service JWT issued with the shared deployment secret
      scheme: bearer
      bearerFormat: JWT